Change variable type of pos in SynthEvent to uint

+++ SharpMIDI.MIDI/DataStorage.cs
@@ -7,8 +7,8 @@ namespace SharpMIDI
     }
     public struct SynthEvent
     {
-        public float pos;
-        public int val;
+        public uint pos;
+        public uint val;
     }
     class MIDIData
     {
